fixed ndisc bug, defined runi interface
There was a typo in NeighBorMod.nc, using a logical negation for clearing a bit instead of a binary negation.
Note: NeighBorMod.nc should be renamed to NeighbourMod.nc (1. syllables are not camelcased. 2. We use BE in the code, so we should not name the files in AE).

Reliable unicast is now defined in RuniC.nc and RuniP.nc.