merged log storage capabilities with rest
merging was quite a pain, so you may want to test it on the hardware. Sending a sens request of type AUTO_SENS will get you a canned sensor data from the log. That can be configured in the appropriate Blink.h enum-definition.

also, you may want to hook this up to the config storage which has been skeletoned already.